# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ThumbBattle API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Thumbnail {
  id: number;
  title: string;
  imageUrl: string;
  channelName: string;
  /**
   * YouTube channelId for YouTube-sourced rows; null for user uploads.
   * @nullable
   */
  channelId?: string | null;
  /**
   * Channel avatar URL (YouTube snippet.thumbnails.medium.url); null for user uploads.
   * @nullable
   */
  channelLogoUrl?: string | null;
  /** One of Gaming, Tutorial, Finance, Music, Lifestyle, Tech, Vlog, Other */
  niche: string;
  /**
   * Click-through rate as a percentage (0-100), null if not provided
   * @nullable
   */
  ctr?: number | null;
  /**
   * Optional public YouTube link to verify the thumbnail
   * @nullable
   */
  youtubeUrl?: string | null;
  /** active or pending */
  status: string;
  wins: number;
  losses: number;
  eloRating: number;
  /**
   * Win rate as a percentage (0-100), null if no battles
   * @nullable
   */
  winRate?: number | null;
  /** Most recent rating snapshots for this thumbnail (chronological,
oldest first, up to 20 points). Embedded so leaderboard
sparklines can render without N+1 per-row fetches. Empty array
for thumbnails that have not battled yet.
 */
  recentRatings?: number[];
  /** Provenance — "user", "youtube", or "seed". */
  source: string;
  /**
   * Source YouTube videoId for thumbnails pulled via trending sync.
   * @nullable
   */
  youtubeVideoId?: string | null;
  /**
   * Latest YouTube view count snapshot (null for non-YouTube rows).
   * @nullable
   */
  viewCount?: number | null;
  /**
   * Views-per-hour (FPH) computed from the delta between the last
two view snapshots. Null until we have at least two snapshots.
This is the "rising" signal — sort by this DESC for trending.

   * @nullable
   */
  viewVelocity?: number | null;
  /**
   * When the source YouTube video was published.
   * @nullable
   */
  publishedAt?: string | null;
  /**
   * Canonical category bucket used for matchmaking and leaderboard
filtering (Blok 5). Mirrors `niche` for user uploads; for
YouTube rows it's the hybrid classifier output. Treat as the
primary bucket; fall back to `niche` only when null (legacy).

   * @nullable
   */
  appCategory?: string | null;
}

export type UploadThumbnailRequestNiche =
  (typeof UploadThumbnailRequestNiche)[keyof typeof UploadThumbnailRequestNiche];

export const UploadThumbnailRequestNiche = {
  Gaming: "Gaming",
  Tutorial: "Tutorial",
  Finance: "Finance",
  Music: "Music",
  Lifestyle: "Lifestyle",
  Tech: "Tech",
  Vlog: "Vlog",
  Other: "Other",
} as const;

export interface UploadThumbnailRequest {
  /**
   * @minLength 1
   * @maxLength 200
   */
  title: string;
  /**
   * @minLength 1
   * @maxLength 120
   */
  channelName: string;
  niche: UploadThumbnailRequestNiche;
  /**
   * Object path returned from /storage/uploads/request-url, or a remote https URL.
   * @minLength 1
   * @maxLength 1000
   */
  imageUrl: string;
  /**
   * Optional CTR percentage (0-100)
   * @minimum 0
   * @maximum 100
   * @nullable
   */
  ctr?: number | null;
  /**
   * Optional public YouTube URL for verification
   * @maxLength 500
   * @nullable
   */
  youtubeUrl?: string | null;
}

export interface BattlePair {
  left: Thumbnail;
  right: Thumbnail;
}

export interface BattleQueueResponse {
  /**
   * @minItems 1
   * @maxItems 10
   */
  pairs: BattlePair[];
}

export interface VoteBody {
  winnerId: number;
  loserId: number;
}

export interface VoteResult {
  winner: Thumbnail;
  loser: Thumbnail;
  totalVotes: number;
}

export interface RecentBattle {
  id: number;
  winnerId: number;
  loserId: number;
  winnerTitle: string;
  loserTitle: string;
  createdAt: string;
}

export interface BattleStats {
  totalVotes: number;
  recentBattles: RecentBattle[];
}

export interface FeedbackBody {
  /**
   * The visitor's feedback message
   * @minLength 1
   * @maxLength 4000
   */
  message: string;
  /**
   * Optional email address so the team can reply
   * @nullable
   */
  email?: string | null;
  /**
   * The URL the visitor was on when sending feedback
   * @nullable
   */
  pageUrl?: string | null;
}

export interface FeedbackResult {
  id: number;
  createdAt: string;
}

export interface WaitlistBody {
  /**
   * Email address to notify when uploads launch
   * @minLength 3
   * @maxLength 254
   */
  email: string;
}

export interface WaitlistResult {
  id: number;
  signedUpAt: string;
}

export interface UploadUrlRequest {
  /**
   * @minLength 1
   * @maxLength 500
   */
  name: string;
  /** @minimum 1 */
  size: number;
  /**
   * @minLength 1
   * @maxLength 200
   */
  contentType: string;
}

export type UploadUrlInfoMetadata = {
  name: string;
  size: number;
  contentType: string;
};

export interface UploadUrlInfo {
  uploadURL: string;
  objectPath: string;
  metadata: UploadUrlInfoMetadata;
}

export interface RatingHistoryPoint {
  rating: number;
  createdAt: string;
}

export interface RatingHistoryResponse {
  thumbnailId: number;
  currentRating: number;
  points: RatingHistoryPoint[];
}

export interface ErrorResponse {
  error: string;
}

export type ListThumbnailsParams = {
  /**
   * Filter to a specific niche (case-insensitive). Omit or "all" returns all niches.
   */
  niche?: string;
  /**
   * Sort order for the rankings. Defaults to elo.
   */
  sort?: ListThumbnailsSort;
};

export type ListThumbnailsSort =
  (typeof ListThumbnailsSort)[keyof typeof ListThumbnailsSort];

export const ListThumbnailsSort = {
  elo: "elo",
  winRate: "winRate",
  ctr: "ctr",
  battles: "battles",
  rising: "rising",
} as const;

export type GetBattlePairParams = {
  /**
   * Restrict the battle pairs to a specific niche.
   */
  niche?: string;
  /**
 * How many pairs to return. Server may return fewer if the niche
does not have enough active thumbnails to satisfy the request.
Always at least 1 on a 200 response.

 * @minimum 1
 * @maximum 10
 */
  count?: number;
};