# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: ThumbBattle API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /thumbnails/battle:
    get:
      operationId: getBattlePair
      tags: [thumbnails]
      summary: Get one or more random battle pairs
      description: |
        Returns a queue of distinct random active thumbnail pairs to compare,
        optionally filtered by niche. Used by the client to maintain a local
        prefetched queue so swipes feel instant. Pass `count` to fetch
        multiple pairs in one round-trip.
      parameters:
        - name: niche
          in: query
          required: false
          schema:
            type: string
          description: Restrict the battle pairs to a specific niche.
        - name: count
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 10
            default: 1
          description: |
            How many pairs to return. Server may return fewer if the niche
            does not have enough active thumbnails to satisfy the request.
            Always at least 1 on a 200 response.
      responses:
        "200":
          description: One or more thumbnail pairs for battles
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BattleQueueResponse"
        "400":
          description: Not enough thumbnails for a battle
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /battles/vote:
    post:
      operationId: castVote
      tags: [battles]
      summary: Cast a vote
      description: Record a vote for a winner against a loser, updates win/loss counts and ELO ratings
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VoteBody"
      responses:
        "200":
          description: Updated thumbnails after vote
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VoteResult"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Thumbnail not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    Thumbnail:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        imageUrl:
          type: string
        channelName:
          type: string
        channelId:
          type: ["string", "null"]
          description: YouTube channelId for YouTube-sourced rows; null for user uploads.
        channelLogoUrl:
          type: ["string", "null"]
          description: Channel avatar URL (YouTube snippet.thumbnails.medium.url); null for user uploads.
        niche:
          type: string
          description: One of Gaming, Tutorial, Finance, Music, Lifestyle, Tech, Vlog, Other
        ctr:
          type: ["number", "null"]
          description: Click-through rate as a percentage (0-100), null if not provided
        youtubeUrl:
          type: ["string", "null"]
          description: Optional public YouTube link to verify the thumbnail
        status:
          type: string
          description: active or pending
        wins:
          type: integer
        losses:
          type: integer
        eloRating:
          type: integer
        winRate:
          type: ["number", "null"]
          description: Win rate as a percentage (0-100), null if no battles
        recentRatings:
          type: array
          items:
            type: integer
          description: |
            Most recent rating snapshots for this thumbnail (chronological,
            oldest first, up to 20 points). Embedded so leaderboard
            sparklines can render without N+1 per-row fetches. Empty array
            for thumbnails that have not battled yet.
        source:
          type: string
          description: Provenance — "user", "youtube", or "seed".
        youtubeVideoId:
          type: ["string", "null"]
          description: Source YouTube videoId for thumbnails pulled via trending sync.
        viewCount:
          type: ["integer", "null"]
          description: Latest YouTube view count snapshot (null for non-YouTube rows).
        viewVelocity:
          type: ["number", "null"]
          description: |
            Views-per-hour (FPH) computed from the delta between the last
            two view snapshots. Null until we have at least two snapshots.
            This is the "rising" signal — sort by this DESC for trending.
        publishedAt:
          type: ["string", "null"]
          format: date-time
          description: When the source YouTube video was published.
        appCategory:
          type: ["string", "null"]
          description: |
            Canonical category bucket used for matchmaking and leaderboard
            filtering (Blok 5). Mirrors `niche` for user uploads; for
            YouTube rows it's the hybrid classifier output. Treat as the
            primary bucket; fall back to `niche` only when null (legacy).
      required:
        - id
        - title
        - imageUrl
        - channelName
        - niche
        - status
        - wins
        - losses
        - eloRating
        - source
...
    BattlePair:
      type: object
      properties:
        left:
          $ref: "#/components/schemas/Thumbnail"
        right:
          $ref: "#/components/schemas/Thumbnail"
      required:
        - left
        - right
    BattleQueueResponse:
      type: object
      properties:
        pairs:
          type: array
          minItems: 1
          maxItems: 10
          items:
            $ref: "#/components/schemas/BattlePair"
      required:
        - pairs
    VoteBody:
      type: object
      properties:
        winnerId:
          type: integer
        loserId:
          type: integer
      required:
        - winnerId
        - loserId
    VoteResult:
      type: object
      properties:
        winner:
          $ref: "#/components/schemas/Thumbnail"
        loser:
          $ref: "#/components/schemas/Thumbnail"
        totalVotes:
          type: integer
      required:
        - winner
        - loser
        - totalVotes
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error