# lib/api-zod/src/generated/types/thumbnail.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ThumbBattle API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Thumbnail {
  id: number;
  title: string;
  imageUrl: string;
  channelName: string;
  /**
   * YouTube channelId for YouTube-sourced rows; null for user uploads.
   * @nullable
   */
  channelId?: string | null;
  /**
   * Channel avatar URL (YouTube snippet.thumbnails.medium.url); null for user uploads.
   * @nullable
   */
  channelLogoUrl?: string | null;
  /** One of Gaming, Tutorial, Finance, Music, Lifestyle, Tech, Vlog, Other */
  niche: string;
  /**
   * Click-through rate as a percentage (0-100), null if not provided
   * @nullable
   */
  ctr?: number | null;
  /**
   * Optional public YouTube link to verify the thumbnail
   * @nullable
   */
  youtubeUrl?: string | null;
  /** active or pending */
  status: string;
  wins: number;
  losses: number;
  eloRating: number;
  /**
   * Win rate as a percentage (0-100), null if no battles
   * @nullable
   */
  winRate?: number | null;
  /** Most recent rating snapshots for this thumbnail (chronological,
oldest first, up to 20 points). Embedded so leaderboard
sparklines can render without N+1 per-row fetches. Empty array
for thumbnails that have not battled yet.
 */
  recentRatings?: number[];
  /** Provenance — "user", "youtube", or "seed". */
  source: string;
  /**
   * Source YouTube videoId for thumbnails pulled via trending sync.
   * @nullable
   */
  youtubeVideoId?: string | null;
  /**
   * Latest YouTube view count snapshot (null for non-YouTube rows).
   * @nullable
   */
  viewCount?: number | null;
  /**
   * Views-per-hour (FPH) computed from the delta between the last
two view snapshots. Null until we have at least two snapshots.
This is the "rising" signal — sort by this DESC for trending.

   * @nullable
   */
  viewVelocity?: number | null;
  /**
   * When the source YouTube video was published.
   * @nullable
   */
  publishedAt?: Date | null;
  /**
   * Canonical category bucket used for matchmaking and leaderboard
filtering (Blok 5). Mirrors `niche` for user uploads; for
YouTube rows it's the hybrid classifier output. Treat as the
primary bucket; fall back to `niche` only when null (legacy).

   * @nullable
   */
  appCategory?: string | null;
}